requirements.txt
asn1crypto==1.3.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.13.18
botocore==1.16.18
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
docutils==0.15.2
Flask==1.1.2
idna==2.9
ijson==2.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
oscrypto==1.2.0
pandas==1.0.3
pycparser==2.20
pycryptodomex==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
shap==0.35.0
six==1.15.0
sklearn==0.0
snowflake-connector-python==2.2.6
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.17
threadpoolctl==2.0.0
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
